Renaming V3 claim card files
