+++ src/applications/claims-status/components/StemClaimListItem.jsx
