+++ src/applications/claims-status/tests/components/StemClaimListItem.unit.spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { expect } from 'chai';
 
-import StemClaimListItem from '../../components/StemClaimListItemV3';
+import StemClaimListItem from '../../components/StemClaimListItem';
 
 describe('<StemClaimListItem>', () => {
   const defaultClaim = {
